refactor: ai 메모 텍스트 네이밍 수정
