+++ apps/web/src/app/wrong/create/components/ai-solution-text/ai-solution-text.css.ts
